Lots of changes
Made the Employee.cs model class again. Changed the isCompleted to a
public variable in the job.cs class. The projectMain.cdproj file needed
to be updated as well to include the InAndOut.cs and saveToDB folder.
Added the SaveToDB folder and the InAndOut.cs class to save and retreive
from the database. Reworked the entire Employeee.xaml  and the
Employee.xaml.cs classes.